pyproject.toml
[project]
name = "etl-pipelines"
version = "1.1.1"
description = "ETL pipelines for Boston housing P1.1.1"
requires-python = ">=3.11"

dependencies = [
  "apache-beam[gcp]>=2.56.0,<3.0.0",
  "google-cloud-bigquery>=3.0.0,<4.0.0",
  "google-cloud-aiplatform>=1.70.0,<2.0.0",
  "kfp>=2.8.0,<3.0.0",
  "google-cloud-pipeline-components>=2.8.0,<3.0.0",
  "pyyaml>=6.0,<7.0",
]

[tool.hatch.build.targets.wheel]
include = [
  "pipelines",
  "dataflow",
  "cleanup",
  "config",
  "scripts",
  "README.md",
  "AGENTS.md",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
